Tests completed

+++ Backend/Total-Knockout-Chess/src/main/java/TotalKnockoutChess/Chess/ChessGameController.java
@@ -1,15 +1,14 @@
 package TotalKnockoutChess.Chess;
 
-import TotalKnockoutChess.Chess.Pieces.Coordinate;
-import TotalKnockoutChess.Lobby.Lobby;
 import TotalKnockoutChess.Lobby.LobbyRepository;
-import TotalKnockoutChess.Users.User;
 import TotalKnockoutChess.Users.UserRepository;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 
 @Api(value = "ChessGameController", description = "Controller used to manage ChessGame entities")
@@ -32,39 +31,46 @@ public class ChessGameController {
 
     @ApiOperation(value = "Creates a new ChessGame instance and saves it to the repository")
     @PostMapping("/{whitePlayer}/{blackPlayer}")
-    public void createChessGame(@PathVariable String whitePlayer, @PathVariable String blackPlayer){
+    public String createChessGame(@PathVariable String whitePlayer, @PathVariable String blackPlayer){
 
         ChessGame game = new ChessGame(whitePlayer, blackPlayer, null);
         chessGameRepository.save(game);
         chessGameRepository.flush();
 
-        game.displayBoard();
-    }
+//        game.displayBoard();
 
-    @ApiOperation(value = "Updates the board if the move is successful")
-    @PutMapping("/{id}/move/{from}/{to}")
-    public String move(@PathVariable int id, @PathVariable String from, @PathVariable String to){
-        ChessGame game = chessGameRepository.findById(id);
-        Coordinate fromCoordinate = Coordinate.fromString(from);
-        Coordinate toCoordinate = Coordinate.fromString(to);
+        // Check if game was successfully saved to database
+        for(ChessGame g : chessGameRepository.findAll()) {
+            // Find exact game match by id
+            if (g.getId() == game.getId()) {
+                return success;
+            }
+        }
 
-        boolean success = game.makeMove(fromCoordinate, toCoordinate);
-
-        // Ensure the board is updated in the database
-        chessGameRepository.flush();
-
-        // Backend text output for debugging
-        game.displayBoard();
-
-        return success + ". Game with id " + game.getId() + " updated";
+        return failure;
     }
 
     @ApiOperation(value = "Deletes a ChessGame entity by the given id")
     @Transactional
     @DeleteMapping("/delete/{id}")
-    public void deleteById(@PathVariable int id){
+    public String deleteById(@PathVariable int id){
 
-        chessGameRepository.deleteById(id);
-        chessGameRepository.flush();
+        // Check for game with matching id
+        for(ChessGame g : chessGameRepository.findAll()){
+            if(g.getId() == id){
+                chessGameRepository.deleteById(id);
+                chessGameRepository.flush();
+
+                return "{\"message\":\"success\"}";
+            }
+        }
+
+        return "{\"message\":\"failure\"}";
+    }
+
+    @ApiOperation(value = "Returns a list of all chess games")
+    @GetMapping(path = "/getAllGames")
+    public List<ChessGame> getAllGames() {
+        return chessGameRepository.findAll();
     }
 }
